Extracted DownloaderFactory to new class
