+++ core/src/main/java/de/danoeh/antennapod/core/service/download/DownloaderFactory.java
@@ -1,0 +1,9 @@
+package de.danoeh.antennapod.core.service.download;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+public interface DownloaderFactory {
+    @Nullable
+    Downloader create(@NonNull DownloadRequest request);
+}
